requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0
django-import-export==2.7.0
et-xmlfile==1.1.0
gunicorn==20.1.0
importlib-metadata==4.9.0
Markdown==3.3.6
markdown2==2.4.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2-binary==2.9.2
python-decouple==3.5
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
